Create tests and new game function
Instead of re-displaying the board every time a game finishes, the
game will not re-display if the game is over. You must start a new game
to get the game to work after the game is over.

I also added tests so I wouldn't have to input a whole game to test.